Rename p1_navigation to udacity-rl
This way I can reuse the repo for other projects without needing to
setup everything again.
