--- p1_navigation/__init__.py
@@ -1,4 +1,0 @@
-from p1_navigation._version import __version__
-from p1_navigation.navigation import cli
-
-name = "p1_navigation"
