+++ udacity_rl/__init__.py
@@ -1,0 +1,4 @@
+from udacity_rl._version import __version__
+from udacity_rl.navigation import cli
+
+name = "udacity_rl"
